PlotVec: add new plot features and tests. Remove old `Traj` code.
