--- stereo/algorithm/trajectory/__init__.py
@@ -1,1 +1,0 @@
-from ._settings import settings
