--- stereo/algorithm/trajectory/_settings.py
@@ -1,76 +1,0 @@
-import logging
-from time import time
-
-from typing import Union, Iterable
-from pathlib import Path
-
-
-class ScanpyConfig:
-    """\
-    Config manager for scanpy.
-    """
-
-    def __init__(
-        self,
-        *,
-        verbosity: str = "warning",
-        plot_suffix: str = "",
-        file_format_data: str = "h5ad",
-        file_format_figs: str = "pdf",
-        autosave: bool = False,
-        autoshow: bool = True,
-        writedir: Union[str, Path] = "./write/",
-        cachedir: Union[str, Path] = "./cache/",
-        datasetdir: Union[str, Path] = "./data/",
-        figdir: Union[str, Path] = "./figures/",
-        cache_compression: Union[str, None] = 'lzf',
-        max_memory=15,
-        n_jobs=1,
-        logfile: Union[str, Path, None] = None,
-        categories_to_ignore: Iterable[str] = ("N/A", "dontknow", "no_gate", "?"),
-        _frameon: bool = True,
-        _vector_friendly: bool = False,
-        _low_resolution_warning: bool = True,
-        n_pcs=50,
-    ):
-        # logging
-        # self._root_logger = _RootLogger(logging.INFO)  # level will be replaced
-        self.logfile = logfile
-        self.verbosity = verbosity
-
-        # rest
-        self.plot_suffix = plot_suffix
-        self.file_format_data = file_format_data
-        self.file_format_figs = file_format_figs
-        self.autosave = autosave
-        self.autoshow = autoshow
-        self.writedir = writedir
-        self.cachedir = cachedir
-        self.datasetdir = datasetdir
-        self.figdir = figdir
-        self.cache_compression = cache_compression
-        self.max_memory = max_memory
-        self.n_jobs = n_jobs
-        self.categories_to_ignore = categories_to_ignore
-        self._frameon = _frameon
-        """bool: See set_figure_params."""
-
-        self._vector_friendly = _vector_friendly
-        """Set to true if you want to include pngs in svgs and pdfs."""
-
-        self._low_resolution_warning = _low_resolution_warning
-        """Print warning when saving a figure with low resolution."""
-
-        self._start = time()
-        """Time when the settings module is first imported."""
-
-        self._previous_time = self._start
-        """Variable for timing program parts."""
-
-        self._previous_memory_usage = -1
-        """Stores the previous memory usage."""
-
-        self.N_PCS = n_pcs
-        """Default number of principal components to use."""
-
-settings = ScanpyConfig()
