--- stereo/algorithm/trajectory/_utils/__init__.py
@@ -1,161 +1,0 @@
-import numpy as np
-from numpy import random
-from typing import Union
-from textwrap import dedent
-
-from stereo.log_manager import logger
-
-# e.g. https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html
-AnyRandom = Union[None, int, random.RandomState]  # maybe in the future random.Generator
-
-
-def get_igraph_from_adjacency(adjacency, directed=None):
-    """Get igraph graph from adjacency matrix."""
-    import igraph as ig
-
-    sources, targets = adjacency.nonzero()
-    weights = adjacency[sources, targets]
-    if isinstance(weights, np.matrix):
-        weights = weights.A1
-    g = ig.Graph(directed=directed)
-    g.add_vertices(adjacency.shape[0])  # this adds adjacency.shape[0] vertices
-    g.add_edges(list(zip(sources, targets)))
-    try:
-        g.es['weight'] = weights
-    except KeyError:
-        pass
-    if g.vcount() != adjacency.shape[0]:
-        logger.warning(
-            f'The constructed graph has only {g.vcount()} nodes. '
-            'Your adjacency matrix contained redundant nodes.'
-        )
-    return g
-
-
-def get_sparse_from_igraph(graph, weight_attr=None):
-    from scipy.sparse import csr_matrix
-
-    edges = graph.get_edgelist()
-    if weight_attr is None:
-        weights = [1] * len(edges)
-    else:
-        weights = graph.es[weight_attr]
-    if not graph.is_directed():
-        edges.extend([(v, u) for u, v in edges])
-        weights.extend(weights)
-    shape = graph.vcount()
-    shape = (shape, shape)
-    if len(edges) > 0:
-        return csr_matrix((weights, zip(*edges)), shape=shape)
-    else:
-        return csr_matrix(shape)
-
-
-# backwards compat... remove this in the future
-def sanitize_anndata(adata):
-    """Transform string annotations to categoricals."""
-    adata._sanitize()
-
-
-def _fallback_to_uns(dct, conns, dists, conns_key, dists_key):
-    if conns is None and conns_key in dct:
-        conns = dct[conns_key]
-    if dists is None and dists_key in dct:
-        dists = dct[dists_key]
-
-    return conns, dists
-
-
-def _doc_params(**kwds):
-    """\
-    Docstrings should start with "\" in the first line for proper formatting.
-    """
-
-    def dec(obj):
-        obj.__orig_doc__ = obj.__doc__
-        obj.__doc__ = dedent(obj.__doc__).format_map(kwds)
-        return obj
-
-    return dec
-
-
-class NeighborsView:
-    """Convenience class for accessing neighbors graph representations.
-
-    Allows to access neighbors distances, connectivities and settings
-    dictionary in a uniform manner.
-
-    Parameters
-    ----------
-
-    adata
-        AnnData object.
-    key
-        This defines where to look for neighbors dictionary,
-        connectivities, distances.
-
-        neigh = NeighborsView(adata, key)
-        neigh['distances']
-        neigh['connectivities']
-        neigh['params']
-        'connectivities' in neigh
-        'params' in neigh
-
-        is the same as
-
-        adata.obsp[adata.uns[key]['distances_key']]
-        adata.obsp[adata.uns[key]['connectivities_key']]
-        adata.uns[key]['params']
-        adata.uns[key]['connectivities_key'] in adata.obsp
-        'params' in adata.uns[key]
-    """
-    def __init__(self, adata, key=None):
-        self._connectivities = None
-        self._distances = None
-
-        if key is None or key == 'neighbors':
-            if 'neighbors' not in adata.uns:
-                raise KeyError('No "neighbors" in .uns')
-            self._neighbors_dict = adata.uns['neighbors']
-            self._conns_key = 'connectivities'
-            self._dists_key = 'distances'
-        else:
-            if key not in adata.uns:
-                raise KeyError(f'No "{key}" in .uns')
-            self._neighbors_dict = adata.uns[key]
-            self._conns_key = self._neighbors_dict['connectivities_key']
-            self._dists_key = self._neighbors_dict['distances_key']
-
-        if self._conns_key in adata.obsp:
-            self._connectivities = adata.obsp[self._conns_key]
-        if self._dists_key in adata.obsp:
-            self._distances = adata.obsp[self._dists_key]
-
-        # fallback to uns
-        self._connectivities, self._distances = _fallback_to_uns(
-            self._neighbors_dict,
-            self._connectivities,
-            self._distances,
-            self._conns_key,
-            self._dists_key,
-        )
-
-    def __getitem__(self, key):
-        if key == 'distances':
-            if 'distances' not in self:
-                raise KeyError(f'No "{self._dists_key}" in .obsp')
-            return self._distances
-        elif key == 'connectivities':
-            if 'connectivities' not in self:
-                raise KeyError(f'No "{self._conns_key}" in .obsp')
-            return self._connectivities
-        else:
-            return self._neighbors_dict[key]
-
-    def __contains__(self, key):
-        if key == 'distances':
-            return self._distances is not None
-        elif key == 'connectivities':
-            return self._connectivities is not None
-        else:
-            return key in self._neighbors_dict
