--- stereo/algorithm/trajectory/tools/_pca.py
@@ -1,1 +1,0 @@
-from ..preprocessing import pca
